added unique and dimensions to products

+++ cart/context_processor.py
@@ -19,7 +19,7 @@ def cart_contents(request):
         cart_items.append({"item_id": item_id, "quantity": quantity, "product": product})
 
     if total < settings.FREE_DELIVERY_THRESHOLD:
-        delivery = total * Decimal(settings.STANDARD_DELIVERY_PERCENTAGE / 100)
+        delivery = total * int(settings.STANDARD_DELIVERY_PERCENTAGE / 100)
         free_delivery_delta = settings.FREE_DELIVERY_THRESHOLD - total
     else:
         delivery = 0
